Support fractional color animation blink rates

+++ src/BaseHmiTypes/Screens/Base/HmiColorAnimation.cs
@@ -4,7 +4,8 @@ public class HmiColorAnimation
 {
     public string? Expression { get; set; }
 
-    public int? BlinkRate { get; set; }
+    /// <summary>Gets or sets the complete blink cycle duration in seconds.</summary>
+    public double? BlinkRate { get; set; }
 
     public List<HmiColorAnimationState> States { get; } = [];
 }
